Updated Individual Income/Questions
We need to decide how we want to do our interactive visualization. We know there is an income problem, but there is also an hours worked problem. It looks like women tend to work 40 hours or less, with some exceptions, while men work more. If we isolate to only part time or only to full time we will lose a lot of data.