Ideas
Maybe we can modify one of these:
https://d3-graph-gallery.com/graph/boxplot_show_individual_points.html

Then we can add a drop down menu. It would be cool if we could also have a histogram when they click it. To see the counts of people in each income bin. Some of it is shocking, but it is more informational that way.